move greetings and safety tests
comment tests

fix import
